Move some es pod consts into the nodespec pkg
