+++ operators/pkg/controller/elasticsearch/pod/pod.go
@@ -5,55 +5,11 @@
 package pod
 
 import (
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/keystore"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/elastic/cloud-on-k8s/operators/pkg/apis/elasticsearch/v1alpha1"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/volume"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/client"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/label"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/network"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/settings"
 )
-
-const (
-	// DefaultImageRepository is the default image name without a tag
-	DefaultImageRepository string = "docker.elastic.co/elasticsearch/elasticsearch"
-
-	// DefaultTerminationGracePeriodSeconds is the termination grace period for the Elasticsearch containers
-	DefaultTerminationGracePeriodSeconds int64 = 120
-)
-
-var (
-	// DefaultContainerPorts are the default Elasticsearch port mappings
-	DefaultContainerPorts = []corev1.ContainerPort{
-		{Name: "http", ContainerPort: network.HTTPPort, Protocol: corev1.ProtocolTCP},
-		{Name: "transport", ContainerPort: network.TransportPort, Protocol: corev1.ProtocolTCP},
-	}
-)
-
-// DefaultAffinity returns the default affinity for pods in a cluster.
-func DefaultAffinity(esName string) *corev1.Affinity {
-	return &corev1.Affinity{
-		// prefer to avoid two pods in the same cluster being co-located on a single node
-		PodAntiAffinity: &corev1.PodAntiAffinity{
-			PreferredDuringSchedulingIgnoredDuringExecution: []corev1.WeightedPodAffinityTerm{
-				{
-					Weight: 100,
-					PodAffinityTerm: corev1.PodAffinityTerm{
-						TopologyKey: "kubernetes.io/hostname",
-						LabelSelector: &metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								label.ClusterNameLabelName: esName,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
 
 // PodWithConfig contains a pod and its configuration
 type PodWithConfig struct {
